Added Gemin support - refactored due to file size
